ai-parking-system/backend/app/api/v1/endpoints/vehicles.py
"""
Vehicle detection and management endpoints
"""

from fastapi import APIRouter, HTTPException, UploadFile, File, Depends
from fastapi.responses import JSONResponse
from typing import List, Dict, Any
import cv2
import numpy as np
import io
from PIL import Image

from app.core.config import settings
from app.services.vehicle_service import VehicleService
from app.schemas.vehicle import VehicleDetectionRequest, VehicleDetectionResponse

router = APIRouter()

@router.post("/detect", response_model=VehicleDetectionResponse)
async def detect_vehicles(
    file: UploadFile = File(...),
    confidence_threshold: float = 0.5
):
    """
    Detect vehicles in uploaded image using AI
    """
    try:
        # Validate file type
        if not file.content_type.startswith('image/'):
            raise HTTPException(status_code=400, detail="File must be an image")
        
        # Read and process image
        image_data = await file.read()
        nparr = np.frombuffer(image_data, np.uint8)
        image = cv2.imdecode(nparr, cv2.IMREAD_COLOR)
        
        if image is None:
            raise HTTPException(status_code=400, detail="Invalid image format")
        
        # Initialize vehicle service
        vehicle_service = VehicleService()
        
        # Detect vehicles
        detections = vehicle_service.detect_vehicles(image, confidence_threshold)
        
        return VehicleDetectionResponse(
            success=True,
            vehicles=detections,
            total_vehicles=len(detections),
            processing_time=0.1  # Placeholder
        )
        
    except Exception as e:
        raise HTTPException(status_code=500, detail=f"Vehicle detection failed: {str(e)}")

@router.get("/types")
async def get_vehicle_types():
    """
    Get supported vehicle types
    """
    return {
        "vehicle_types": ["car", "truck", "motorcycle", "bus", "van"],
        "description": "Supported vehicle types for detection"
    }

@router.get("/stats")
async def get_detection_stats():
    """
    Get vehicle detection statistics
    """
    return {
        "total_detections": 0,  # Placeholder
        "accuracy": 0.95,
        "processing_time_avg": 0.1
    }
