ai-parking-system/requirements.txt
# Core ML and Computer Vision
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
ultralytics>=8.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# OCR and Image Processing
pytesseract>=0.3.10
Pillow>=10.0.0
imutils>=0.5.4

# Backend Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.0
redis>=5.0.0
celery>=5.3.0

# Authentication and Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# API and Serialization
requests>=2.31.0
httpx>=0.25.0
pydantic-settings>=2.1.0

# Database
asyncpg>=0.29.0
aioredis>=2.0.0

# Monitoring and Logging
prometheus-client>=0.19.0
structlog>=23.2.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Development
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.7.0

# Utilities
python-dotenv>=1.0.0
click>=8.1.0
rich>=13.0.0
